src/splitters/text_splitter.py
from langchain_text_splitters import RecursiveCharacterTextSplitter

def split_docs(documents, chunk_size, chunk_overlap):
    splitter = RecursiveCharacterTextSplitter(
        chunk_size=chunk_size,
        chunk_overlap=chunk_overlap
    )
    return splitter.split_documents(documents)
