requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
chromadb
pypdf
python-dotenv
